쿼리DSL 적용 후 랜덤한 음식들을 in절로 가져옴

+++ src/main/java/com/etw/etwapi/common/RandomNum.java
@@ -4,12 +4,12 @@ import java.util.*;
 
 public class RandomNum {
 
-    public static Set<Integer> getRandomSet(int size, int foodSize){
-        Set<Integer> randomSet = new HashSet<>();
+    public static Set<Long> getRandomSet(int size, int foodSize){
+        Set<Long> randomSet = new HashSet<>();
         Random random = new Random();
 
         while(randomSet.size() != size){
-            randomSet.add(random.nextInt(foodSize)+1);
+            randomSet.add((long) (random.nextInt(foodSize)+1));
         }
 
         return randomSet;
